Drop c_64 and c_32
