+++ mjit_compiler.rb
@@ -2,18 +2,13 @@
 # TODO: Merge this to mjit.rb
 if RubyVM::MJIT.enabled?
   begin
-    require 'etc'
     require 'fiddle'
+    require 'fiddle/import'
   rescue LoadError
     return # skip miniruby
   end
 
-  if Fiddle::SIZEOF_VOIDP == 8
-    require 'mjit/c_64'
-  else
-    require 'mjit/c_32'
-  end
-
+  require "mjit/c_type"
   require "mjit/instruction"
   require "mjit/compiler"
 end
